[client,sdl] move functions to proper place
